rename session initializer module
