+++ client/src/lib.rs
@@ -28,8 +28,8 @@ mod _metrics;
 mod _network_view;
 mod _raw_session;
 mod _routing_session;
+mod _session_initializer;
 mod _session_layer;
-mod _session_protocol;
 mod _session_state;
 mod _session_traits;
 mod _tcp_registry;
