+++ client/src/testing/accessors.rs
@@ -1,7 +1,7 @@
 use futures::future::LocalBoxFuture;
 use std::net::SocketAddr;
 
-use crate::_session_protocol::SessionInitializer;
+use crate::_session_initializer::SessionInitializer;
 
 /// Give access to private fields for testing purposes.
 pub trait SessionLayerPrivate {
